feat: Update transaction amount type to int
The code changes modify the `Transaction` struct and `TransactionCreateRequest` struct to change the `Amount` field type from `float64` to `int`. This update ensures that the transaction amount is represented as an integer value.

Recent repository commits indicate a convention of using a prefix to indicate the type of change, such as "feat" for feature additions and "fix" for bug fixes.

+++ internal/domain/transaction.go
@@ -11,17 +11,17 @@ type Transaction struct {
 	CreatedAt   time.Time
 	UpdatedAt   time.Time
 	DeletedAt   gorm.DeletedAt `gorm:"index"`
-	Amount      float64
+	Amount      int
 	PayerID     int //User ID
 	RecieverID  int //User ID
 	Description string
 }
 
 type TransactionCreateRequest struct {
-	Amount      float64 `json:"amount"`
-	PayerID     int     `json:"payer_id"`
-	RecieverID  int     `json:"reciever_id"`
-	Description string  `json:"description"`
+	Amount      int    `json:"amount"`
+	PayerID     int    `json:"payer_id"`
+	RecieverID  int    `json:"reciever_id"`
+	Description string `json:"description"`
 }
 
 type TransactionPageRequest struct {
@@ -32,7 +32,7 @@ type TransactionPageRequest struct {
 }
 
 type TransactionPageResponse struct {
-	Amount      float64 `json:"amount"`
+	Amount      int `json:"amount"`
 	PayerID     int     `json:"payer_id"`
 	RecieverID  int     `json:"reciever_id"`
 	Description string  `json:"description"`
